refactor(spectrogram): windowed rendering as a strategy; spectrogram-windowed becomes a shim
- New src/plugins/spectrogram-windowing.ts: SegmentManager (segment map, farthest-first
  eviction, uncovered-range detection, progressive-load scheduling via an injected
  Scope-shaped timer) plus the segment-canvas renderer and DOM/zoom helpers extracted from
  the old WindowedSpectrogramPlugin. Host-agnostic (deps-injected), unit-tested directly
  (spectrogram-windowing.test.ts, 37 tests).
- New src/spectrogram-setup.ts: the actual spectrogramSetup()/validateOptions()
  implementation, moved out of src/plugins/ so it can be a real named export consumed by both
  entry points without breaking rollup's single-default-export CJS/UMD builds for
  src/plugins/*.ts. spectrogram.ts adds `rendering: 'full' | 'windowed'` plus the windowed-only
  options (windowSize, bufferSize, progressiveLoading) and dispatches to a SegmentManager when
  rendering: 'windowed'; the worker-timeout, worker lifecycle, and computeFrequencies path are
  now shared by both rendering modes (unifying the Phase-1 hard-coded 30s windowed timeout onto
  the configurable workerTimeout option).
- spectrogram-windowed.ts is now a deprecated, definePlugin-based shim that delegates straight
  into spectrogramSetup() with rendering forced to 'windowed' and its (unchanged) option names
  mapped through; a thin wrapper class preserves create-time validateOptions() throws.
- rollup.config.js: exclude spectrogram-windowing.ts from the per-file plugin-bundling glob
  (it has no default export and isn't a public plugin entry).
- Test adaptations: spectrogram-windowed-destroy.test.ts and the windowed-only blocks in
  spectrogram-praat-options.test.ts/spectrogram-worker-errors.test.ts now call _init() and
  reach windowed internals via __spectrogramInternalsForTests().windowed, mirroring Task 2's
  __spectrogramInternalsForTests() precedent for the merged plugin.
