+++ src/__tests__/spectrogram-windowed-destroy.test.ts
@@ -15,6 +15,37 @@ jest.mock(
 
 import WindowedSpectrogramPlugin from '../plugins/spectrogram-windowed.js'
 
+// WindowedSpectrogramPlugin is now a thin shim that delegates into spectrogram.ts's
+// definePlugin() setup (see spectrogram-windowed.ts's doc comment) - its Api (including the
+// test-only __spectrogramInternalsForTests() hatch) only exists on the instance once _init()
+// has run, not at .create() time. Same _init() precedent as spectrogram-destroy.test.ts; see
+// spectrogram.ts's WindowedTestInternals type for what's exposed under .windowed here.
+//
+// __windowedInternals() below is the private-poke adaptation for this file (Phase 4 Task 3):
+// every direct instance-field poke the pre-unification version of this test used
+// ((plugin as any).segments, .maxRetainedSegments, .buffer, .evictDistantSegments(), etc.) is
+// replaced with the equivalent read through __spectrogramInternalsForTests().windowed, whose
+// segmentManager is a real SegmentManager class instance - its own methods call each other via
+// `this.foo()`, so jest.spyOn(internals.segmentManager, 'foo') still correctly intercepts
+// internal calls (unlike spying on a closure-merged Api field - see Task 2's report).
+function createFakeWaveSurfer(overrides: Record<string, unknown> = {}) {
+  const wrapper = document.createElement('div')
+  Object.defineProperty(wrapper, 'offsetWidth', { value: 600, configurable: true })
+  Object.defineProperty(wrapper, 'clientWidth', { value: 600, configurable: true })
+  return {
+    options: {},
+    getWrapper: () => wrapper,
+    getDecodedData: () => null,
+    on: () => () => undefined,
+    ...overrides,
+  }
+}
+
+function initPlugin(plugin: any) {
+  plugin._init(createFakeWaveSurfer() as any)
+  return plugin.__spectrogramInternalsForTests().windowed
+}
+
 describe('WindowedSpectrogramPlugin destroy', () => {
   it('delivers the destroy event to subscribers', () => {
     const plugin = WindowedSpectrogramPlugin.create({})
@@ -35,11 +66,13 @@ describe('WindowedSpectrogramPlugin destroy', () => {
 
   describe('segment eviction', () => {
     it('caps retained segments', () => {
-      const plugin = WindowedSpectrogramPlugin.create({})
-      const cap = (plugin as any).maxRetainedSegments
+      const plugin: any = WindowedSpectrogramPlugin.create({})
+      const internals = initPlugin(plugin)
+      const manager = internals.segmentManager
+      const cap = manager.maxRetainedSegments
       expect(typeof cap).toBe('number')
       for (let i = 0; i < cap + 10; i++) {
-        ;(plugin as any).segments.set(i, {
+        manager.segments.set(i, {
           startTime: i * 30,
           endTime: (i + 1) * 30,
           startPixel: i * 30,
@@ -48,13 +81,15 @@ describe('WindowedSpectrogramPlugin destroy', () => {
           frequencies: [],
         })
       }
-      ;(plugin as any).evictDistantSegments(0) // current view at t=0
-      expect((plugin as any).segments.size).toBeLessThanOrEqual(cap)
+      manager.evictDistantSegments(0) // current view at t=0
+      expect(manager.segments.size).toBeLessThanOrEqual(cap)
     })
 
     it('evicts segments farthest from the current time and removes their canvases from the DOM', () => {
-      const plugin = WindowedSpectrogramPlugin.create({})
-      const cap = (plugin as any).maxRetainedSegments
+      const plugin: any = WindowedSpectrogramPlugin.create({})
+      const internals = initPlugin(plugin)
+      const manager = internals.segmentManager
+      const cap = manager.maxRetainedSegments
       const container = document.createElement('div')
       document.body.appendChild(container)
 
@@ -62,7 +97,7 @@ describe('WindowedSpectrogramPlugin destroy', () => {
       for (let i = 0; i < total; i++) {
         const canvas = document.createElement('canvas')
         container.appendChild(canvas)
-        ;(plugin as any).segments.set(i, {
+        manager.segments.set(i, {
           startTime: i * 30,
           endTime: (i + 1) * 30,
           startPixel: i * 30,
@@ -74,36 +109,38 @@ describe('WindowedSpectrogramPlugin destroy', () => {
 
       // Current view is near the start (t=0), so the segments with the highest
       // indices (farthest midpoints) should be the ones evicted.
-      ;(plugin as any).evictDistantSegments(0)
+      manager.evictDistantSegments(0)
 
-      expect((plugin as any).segments.size).toBe(cap)
+      expect(manager.segments.size).toBe(cap)
       // Every remaining segment's canvas should still be attached to the DOM.
-      for (const segment of (plugin as any).segments.values()) {
+      for (const segment of manager.segments.values()) {
         expect(segment.canvas.isConnected).toBe(true)
       }
       // The DOM should have exactly as many canvases left as retained segments -
       // the evicted ones were actually removed, not just dropped from the map.
       expect(container.children.length).toBe(cap)
       // The segments closest to t=0 (lowest indices) must have survived.
-      expect((plugin as any).segments.has(0)).toBe(true)
-      expect((plugin as any).segments.has(total - 1)).toBe(false)
+      expect(manager.segments.has(0)).toBe(true)
+      expect(manager.segments.has(total - 1)).toBe(false)
     })
 
     it('enforces the segment cap during progressive loading even with no renderVisibleWindow calls', async () => {
       jest.useFakeTimers()
       try {
-        const plugin = WindowedSpectrogramPlugin.create({ progressiveLoading: true })
-        ;(plugin as any).maxRetainedSegments = 3
+        const plugin: any = WindowedSpectrogramPlugin.create({ progressiveLoading: true })
+        const internals = initPlugin(plugin)
+        const manager = internals.segmentManager
+        manager.maxRetainedSegments = 3
         // Long enough for far more than maxRetainedSegments 30s segments.
-        ;(plugin as any).buffer = { duration: 30 * 20 }
-        ;(plugin as any).isProgressiveLoading = true
+        plugin.__spectrogramInternalsForTests().buffer = { duration: 30 * 20 }
+        manager.isProgressiveLoading = true
 
         // Stand in for the real FFT pipeline: just record a segment for the requested range,
         // the same way generateSegments would after computing frequencies.
-        jest.spyOn(plugin as any, 'generateSegments').mockImplementation(async (...args: unknown[]) => {
+        jest.spyOn(manager, 'generateSegments').mockImplementation(async (...args: unknown[]) => {
           const [start, end] = args as [number, number]
           const key = `${Math.floor(start * 10)}_${Math.floor(end * 10)}`
-          ;(plugin as any).segments.set(key, {
+          manager.segments.set(key, {
             startTime: start,
             endTime: end,
             startPixel: 0,
@@ -116,10 +153,10 @@ describe('WindowedSpectrogramPlugin destroy', () => {
         // Drive several progressive-load ticks directly (no scroll/resize/playback, so
         // renderVisibleWindow - and its evictDistantSegments call - never runs).
         for (let i = 0; i < 10; i++) {
-          await (plugin as any).progressiveLoadNextSegment()
+          await manager.progressiveLoadNextSegment()
         }
 
-        expect((plugin as any).segments.size).toBeLessThanOrEqual(3)
+        expect(manager.segments.size).toBeLessThanOrEqual(3)
       } finally {
         jest.useRealTimers()
       }
@@ -130,17 +167,19 @@ describe('WindowedSpectrogramPlugin destroy', () => {
     it('does not re-arm the progressive-load timer once destroy runs mid-await', async () => {
       jest.useFakeTimers()
       try {
-        const plugin = WindowedSpectrogramPlugin.create({ progressiveLoading: true })
-        ;(plugin as any).buffer = { duration: 120 }
-        ;(plugin as any).isProgressiveLoading = true
+        const plugin: any = WindowedSpectrogramPlugin.create({ progressiveLoading: true })
+        const internals = initPlugin(plugin)
+        const manager = internals.segmentManager
+        plugin.__spectrogramInternalsForTests().buffer = { duration: 120 }
+        manager.isProgressiveLoading = true
 
         let resolveGenerate: () => void = () => {}
         const generatePromise = new Promise<void>((resolve) => {
           resolveGenerate = resolve
         })
-        jest.spyOn(plugin as any, 'generateSegments').mockReturnValue(generatePromise)
+        jest.spyOn(manager, 'generateSegments').mockReturnValue(generatePromise)
 
-        const continuation = (plugin as any).progressiveLoadNextSegment()
+        const continuation = manager.progressiveLoadNextSegment()
 
         // Destroy while the generateSegments() await is still pending.
         plugin.destroy()
@@ -159,19 +198,23 @@ describe('WindowedSpectrogramPlugin destroy', () => {
     })
 
     it('does not throw or touch destroyed DOM state when renderVisibleWindow resumes after destroy', async () => {
-      const plugin = WindowedSpectrogramPlugin.create({})
-      ;(plugin as any).buffer = { duration: 60 }
+      const plugin: any = WindowedSpectrogramPlugin.create({})
       const fakeWrapper = document.createElement('div')
-      ;(plugin as any).wavesurfer = { getWrapper: () => fakeWrapper, options: {} }
+      Object.defineProperty(fakeWrapper, 'offsetWidth', { value: 600, configurable: true })
+      Object.defineProperty(fakeWrapper, 'clientWidth', { value: 600, configurable: true })
+      plugin._init(createFakeWaveSurfer({ getWrapper: () => fakeWrapper }) as any)
+      const internals = plugin.__spectrogramInternalsForTests().windowed
+      const manager = internals.segmentManager
+      plugin.__spectrogramInternalsForTests().buffer = { duration: 60 }
 
       let resolveGenerate: () => void = () => {}
       const generatePromise = new Promise<void>((resolve) => {
         resolveGenerate = resolve
       })
-      jest.spyOn(plugin as any, 'generateSegments').mockReturnValue(generatePromise)
-      const evictSpy = jest.spyOn(plugin as any, 'evictDistantSegments')
+      jest.spyOn(manager, 'generateSegments').mockReturnValue(generatePromise)
+      const evictSpy = jest.spyOn(manager, 'evictDistantSegments')
 
-      const renderPromise = (plugin as any).renderVisibleWindow()
+      const renderPromise = manager.renderVisibleWindow()
       plugin.destroy()
       resolveGenerate()
